Refactor helper functions for map conversion and merging.
Moved `StructToMap` and related helper functions to the `charts` package for better modularity and reusability. Updated references in tests to use the refactored functions from the `charts` package. This improves code organization and reduces redundancy.
